+++ Samples~/SDK Demo/Scripts/ConsentTestScene.cs
@@ -14,6 +14,7 @@ namespace BidscubeSDK.Controllers
         [Header("SDK Configuration")]
         [SerializeField] private string _placementId = "20212";
         [SerializeField] private string _baseURL = Constants.BaseURL;
+        [SerializeField] private string _adRequestAuthority = "";
         [SerializeField] private bool _enableDebugMode = true;
         [SerializeField] private bool _enableLogging = true;
 
@@ -107,13 +108,13 @@ namespace BidscubeSDK.Controllers
         {
             LogMessage("Initializing Bidscube SDK...");
 
-            var config = new SDKConfig.Builder()
+            var builder = new SDKConfig.Builder()
                 .EnableLogging(_enableLogging)
                 .EnableDebugMode(_enableDebugMode)
-                .BaseURL(_baseURL)
                 .DefaultAdTimeout(30000)
-                .DefaultAdPosition(AdPosition.Unknown)
-                .Build();
+                .DefaultAdPosition(AdPosition.Unknown);
+            SampleSspAuthorityConfig.ApplyTo(builder, _baseURL, _adRequestAuthority);
+            var config = builder.Build();
 
             BidscubeSDK.Initialize(config);